% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_crane.R
\name{read_crane_file}
\alias{read_crane_file}
\title{Read Crane File}
\usage{
read_crane_file(file_path, include_angle = FALSE)
}
\arguments{
\item{file_path}{A file path string, which is the file to read}

\item{include_angle}{A boolean, which indicates if the file should have angle measurements}
}
\value{
a data frame with columns "time"; "torque"; "angle" (optional)
}
\description{
This function takes a file path, looks for .csv files and reads in an expected crane file format
}
\examples{

read_crane_file("Path/to/filename.csv")
read_crane_file("Path/to/filename.csv", include_angle = TRUE)
}
