% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_crane.R
\name{read_crane_folder}
\alias{read_crane_folder}
\title{Read Crane Folder}
\usage{
read_crane_folder(folder_path, include_angle = FALSE)
}
\arguments{
\item{folder_path}{A folder path string, which contains files (or subfolders with files)}

\item{include_angle}{A boolean, which indicates if the file should have angle measurements}
}
\value{
a data frame with columns "path"; time"; "torque"; "angle" (optional)
}
\description{
This function takes a path to a folder containing crane files, and recursively reads all data by calling read_crane_file()
}
\examples{
read_crane_folder("Path/to/folder")
read_crane_folder("Path/to/folder", include_angle = TRUE)

}
